+++ grngame/platform/virtual_memory.h
@@ -1,18 +1,19 @@
 #pragma once
 #include <stdbool.h>
 #include <stddef.h>
+#include "grngame/math/types.h"
 
 /// reserves a contiguous virtual address range without allocating RAM
-void *VirtualMemoryReserve(size_t size);
+void *VirtualMemoryReserve(uint64 size);
 
 /// commits pages in [ptr, ptr+size[, backing them with real RAM
-bool VirtualMemoryCommit(void *ptr, size_t size);
+bool VirtualMemoryCommit(void *ptr, uint64 size);
 
 /// decommits pages, freeing RAM but keeping the address range reserved
-bool VirtualMemoryDecommit(void *ptr, size_t size);
+bool VirtualMemoryDecommit(void *ptr, uint64 size);
 
 /// releases the entire reserved range, ptr must be the original reserve base
-bool VirtualMemoryRelease(void *ptr, size_t size);
+bool VirtualMemoryRelease(void *ptr, uint64 size);
 
 /// Page size
-size_t VirtualMemoryPageSize();
+uint64 VirtualMemoryPageSize();
