+++ grngame/utils/simd.h
@@ -7,15 +7,15 @@ BEGIN_DECLARATIONS;
 
 typedef struct
 {
-    uint8_t r[256];
-    uint8_t g[256];
-    uint8_t b[256];
+    uint8 r[256];
+    uint8 g[256];
+    uint8 b[256];
     int32 count;
 } PaletteSIMD;
 
 void BuildPaletteSIMD(PaletteSIMD *SDL_RESTRICT out, const SDL_Color *SDL_RESTRICT colors, int32 count);
 
-void RemapImagePalette(uint8_t *SDL_RESTRICT pixels, int32_t w, int32_t h, int32_t pitch,
-                       const PaletteSIMD *SDL_RESTRICT pal, const uint8_t alpha_lut[256]);
+void RemapImagePalette(uint8 *SDL_RESTRICT pixels, uint16 w, uint16 h, uint16 pitch,
+                       const PaletteSIMD *SDL_RESTRICT pal, const uint8 alpha_lut[256]);
 
 END_DECLARATIONS;
